Catch sync storage write errors
Sync storage has relatively low limits on how much data can be stored,
so errors are likely to be thrown here with extensive use of Drumline. I
decided to keep the blocked domains list in sync storage rather than
local storage because I and most users will probably always prefer their
blocked domains list to be reliably saved rather than having to
regularly export it or lose it since local storage is cleared when the
extension is uninstalled. The fact that (at least for now) Drumline only
allows domains to be blocked and not specific URLs helps to keep the
data under the limit.